processing/router_module/requirements.txt
# WaddleBot Flask Module - Dependencies

# Core Framework
quart>=0.20.0
hypercorn>=0.16.0

# gRPC Support
grpcio>=1.67.0,<2.0.0
grpcio-tools>=1.67.0,<2.0.0

# HTTP Client (async)
httpx>=0.27.0,<0.28.0
aiohttp>=3.12.14,<4.0.0

# Environment
python-dotenv>=1.0.0

# Translation
googletrans-py>=4.0.0
google-cloud-translate>=3.12.0  # Optional, for premium Google Cloud API
cachetools>=5.3.0  # For in-memory LRU cache

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
